+++ actionpack/test/dispatch/middleware_stack_test.rb
@@ -9,6 +9,7 @@ class MiddlewareStackTest < ActiveSupport::TestCase
     end
 
     def call(env)
+      env[:middleware_run] << self.class if env.key?(:middleware_run)
       @app.call(env)
     end
   end
@@ -212,5 +213,63 @@ class MiddlewareStackTest < ActiveSupport::TestCase
 
   test "includes a middleware" do
     assert_equal true, @stack.include?(ActionDispatch::MiddlewareStack::Middleware.new(BarMiddleware, nil, nil))
+  end
+
+  test "can add another middleware stack to a stack" do
+    nested_stack = ActionDispatch::MiddlewareStack.new
+    nested_stack.use BazMiddleware
+    nested_stack.use HiyaMiddleware
+    @stack.insert_before BarMiddleware, nested_stack
+
+    assert @stack.include?(nested_stack)
+  end
+
+  test "includes nested stacks when building" do
+    nested_stack = ActionDispatch::MiddlewareStack.new
+    nested_stack.use BazMiddleware
+    @stack.insert_before BarMiddleware, nested_stack
+    nested_stack.use HiyaMiddleware
+
+    app = @stack.build(proc { |env| [200, {}, []] })
+    env = { middleware_run: [] }
+    app.call(env)
+
+    assert_equal [FooMiddleware, BazMiddleware, HiyaMiddleware, BarMiddleware], env[:middleware_run]
+  end
+
+  test "can nest multiple levels" do
+    outer_nested_stack = ActionDispatch::MiddlewareStack.new
+    inner_nested_stack = ActionDispatch::MiddlewareStack.new
+    outer_nested_stack.use BazMiddleware
+    outer_nested_stack.use inner_nested_stack
+    inner_nested_stack.use HiyaMiddleware
+    @stack.insert_before BarMiddleware, outer_nested_stack
+
+    app = @stack.build(proc { |env| [200, {}, []] })
+    env = { middleware_run: [] }
+    app.call(env)
+
+    assert_equal [FooMiddleware, BazMiddleware, HiyaMiddleware, BarMiddleware], env[:middleware_run]
+  end
+
+  test "includes nested stacks when building with instrumentation" do
+    nested_stack = ActionDispatch::MiddlewareStack.new
+    nested_stack.use BazMiddleware
+    @stack.insert_before BarMiddleware, nested_stack
+    nested_stack.use HiyaMiddleware
+
+    events = []
+
+    subscriber = proc do |*args|
+      events << ActiveSupport::Notifications::Event.new(*args)
+    end
+
+    env = { middleware_run: [] }
+    ActiveSupport::Notifications.subscribed(subscriber, "process_middleware.action_dispatch") do
+      app = @stack.build(proc { |env| [200, {}, []] })
+      app.call(env)
+    end
+
+    assert_equal [FooMiddleware, BazMiddleware, HiyaMiddleware, BarMiddleware], env[:middleware_run]
   end
 end